--- AMiManera/AMiManeraPersistencia/src/main/java/com/restaurante/amm/amimaneracapadominioda/IMesaDAO.java
@@ -1,7 +1,0 @@
-package com.restaurante.amm.amimaneracapadominioda;
-
-import com.restaurante.amm.amimaneracapadominiomodel.gestionmesas.Mesa;
-
-public interface IMesaDAO extends ICrud<Mesa>{
-    
-}
