+++ AMiManera/AMiManeraPersistencia/src/main/java/com/restaurante/amm/amimanerapersistencia/dao/PersonaJuridicaDAOImpl.java
@@ -1,4 +1,4 @@
-package com.restaurante.amm.amimaneracapadominioda.mysql;
+package com.restaurante.amm.amimanerapersistencia.dao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -8,22 +8,24 @@ import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.restaurante.amm.amimaneracapadominiomodel.clientes.PersonaNatural;
-import com.restaurante.amm.amimaneracapadominioda.IPersonaNaturalDAO;
+import com.restaurante.amm.amimaneracapadominiomodel.clientes.PersonaJuridica;
+import com.restaurante.amm.amimanerapersistencia.daoimpl.IPersonaJuridicaDAO;
 import com.restaurante.amm.amimaneradbmanager.DBManager;
 
-public class PersonaNaturalDAOImpl implements IPersonaNaturalDAO{
+public class PersonaJuridicaDAOImpl implements IPersonaJuridicaDAO{
     @Override
-    public int insertar(PersonaNatural personaNatural){
-        String sql="INSERT PersonaNatural (idCliente,nombre,telefono,correo,"
-                + "DNI) VALUES (?,?,?,?,?)";
+    public int insertar(PersonaJuridica personaJuridica){
+        String sql="INSERT PersonaJuridica(idCliente,nombre,telefono,correo,razonSocial,"
+                + "RUC) VALUES (?,?,?,?,?,?)";
         try(Connection conn=DBManager.getInstance().getConnection();
                 PreparedStatement ps=conn.prepareStatement(sql,Statement.RETURN_GENERATED_KEYS)){
-            ps.setInt(1,personaNatural.getIdCliente());
-            ps.setString(2,personaNatural.getNombre());
-            ps.setInt(3, personaNatural.getTelefono());
-            ps.setString(4, personaNatural.getCorreo());
-            ps.setString(5, personaNatural.getDNI());
+            
+            ps.setInt(1,personaJuridica.getIdCliente());
+            ps.setString(2,personaJuridica.getNombre());
+            ps.setInt(3, personaJuridica.getTelefono());
+            ps.setString(4, personaJuridica.getCorreo());
+            ps.setString(5, personaJuridica.getRazonSocial());
+            ps.setString(6, personaJuridica.getRUC());
             
             if(ps.executeUpdate()==0){
                 System.err.println("El registro no se inserto.");
@@ -44,16 +46,18 @@ public class PersonaNaturalDAOImpl implements IPersonaNaturalDAO{
     }
     
     @Override
-    public boolean modificar(PersonaNatural personaNatural){
-        String sql="UPDATE PersonaNatural SET nombre=?,telefono=?,correo=?,"
-                + "DNI=? WHERE idCliente=?";
+    public boolean modificar(PersonaJuridica personaJuridica){
+        String sql="UPDATE PersonaJuridica SET nombre=?,telefono=?,correo=?,"
+                + "razonSocial=?,RUC=? WHERE idCliente=?";
         try(Connection conn=DBManager.getInstance().getConnection();
                 PreparedStatement ps=conn.prepareStatement(sql)){
-            ps.setInt(1,personaNatural.getIdCliente());
-            ps.setString(2,personaNatural.getNombre());
-            ps.setInt(3, personaNatural.getTelefono());
-            ps.setString(4, personaNatural.getCorreo());
-            ps.setString(5, personaNatural.getDNI());
+            
+            ps.setInt(1,personaJuridica.getIdCliente());
+            ps.setString(2,personaJuridica.getNombre());
+            ps.setInt(3, personaJuridica.getTelefono());
+            ps.setString(4, personaJuridica.getCorreo());
+            ps.setString(5, personaJuridica.getRazonSocial());
+            ps.setString(6, personaJuridica.getRUC());
             
             return ps.executeUpdate()>0;
         }
@@ -70,7 +74,7 @@ public class PersonaNaturalDAOImpl implements IPersonaNaturalDAO{
     
     @Override
     public boolean eliminar(int id){
-        String sql="DELETE FROM PersonaNatural WHERE idCliente=?";
+        String sql="DELETE FROM PersonaJuridica WHERE idCliente=?";
         try (Connection conn = DBManager.getInstance().getConnection();
             PreparedStatement ps = conn.prepareStatement(sql)) {
             
@@ -89,8 +93,8 @@ public class PersonaNaturalDAOImpl implements IPersonaNaturalDAO{
     }
     
     @Override
-    public PersonaNatural buscar(int id){
-        String sql="SELECT * FROM PersonaNatural WHERE idCliente=?";
+    public PersonaJuridica buscar(int id){
+        String sql="SELECT * FROM PersonaJuridica WHERE idCliente=?";
         try (Connection conn = DBManager.getInstance().getConnection();
             PreparedStatement ps = conn.prepareStatement(sql)) {
             
@@ -102,13 +106,15 @@ public class PersonaNaturalDAOImpl implements IPersonaNaturalDAO{
                     return null;
                 }
 
-                PersonaNatural personaNatural = new PersonaNatural();
-                personaNatural.setIdCliente(rs.getInt("idCliente"));
-                personaNatural.setNombre(rs.getString("nombre"));
-                personaNatural.setTelefono(rs.getInt("telefono"));
-                personaNatural.setCorreo(rs.getString("correo"));
-                personaNatural.setDNI(rs.getString("DNI"));
-                return personaNatural;
+                PersonaJuridica personaJuridica = new PersonaJuridica();
+                personaJuridica.setIdCliente(rs.getInt("idCliente"));
+                personaJuridica.setNombre(rs.getString("nombre"));
+                personaJuridica.setTelefono(rs.getInt("telefono"));
+                personaJuridica.setCorreo(rs.getString("correo"));
+                personaJuridica.setRazonSocial(rs.getString("razonSocial"));
+                personaJuridica.setRUC(rs.getString("RUC"));
+                
+                return personaJuridica;
             }     
         }        
         catch (SQLException e) {
@@ -122,20 +128,22 @@ public class PersonaNaturalDAOImpl implements IPersonaNaturalDAO{
     }
     
     @Override
-    public List<PersonaNatural> listar(){
-        String sql="SELECT * FROM PersonaNatural";
+    public List<PersonaJuridica> listar(){
+        String sql="SELECT * FROM PersonaJuridica";
         try (Connection conn = DBManager.getInstance().getConnection();
                 PreparedStatement ps = conn.prepareStatement(sql);
                 ResultSet rs=ps.executeQuery()) {
             
-            List<PersonaNatural> clientes = new ArrayList<>();
+            List<PersonaJuridica> clientes = new ArrayList<>();
             while (rs.next()) {
-                PersonaNatural cliente = new PersonaNatural();
+                PersonaJuridica cliente = new PersonaJuridica();
                 cliente.setIdCliente(rs.getInt("idCliente"));
                 cliente.setNombre(rs.getString("nombre"));
                 cliente.setTelefono(rs.getInt("telefono"));
                 cliente.setCorreo(rs.getString("correo"));
-                cliente.setDNI(rs.getString("DNI"));
+                cliente.setRazonSocial(rs.getString("razonSocial"));
+                cliente.setRUC(rs.getString("RUC"));
+                
                 clientes.add(cliente);
             }
             return clientes;
