+++ AMiManera/AMiManeraPersistencia/src/main/java/com/restaurante/amm/amimanerapersistencia/daoimpl/IMesaDAO.java
@@ -1,0 +1,7 @@
+package com.restaurante.amm.amimanerapersistencia.daoimpl;
+
+import com.restaurante.amm.amimaneracapadominiomodel.gestionmesas.Mesa;
+
+public interface IMesaDAO extends ICrud<Mesa> {
+    
+}
