+++ AMiManera/AMiManeraPersistencia/src/main/java/com/restaurante/amm/amimanerapersistencia/daoimpl/IPedidoDAO.java
@@ -1,4 +1,4 @@
-package com.restaurante.amm.amimaneracapadominioda;
+package com.restaurante.amm.amimanerapersistencia.daoimpl;
 
 import com.restaurante.amm.amimaneracapadominiomodel.pedidos.Pedido;
